fix(log_view): add medication button
TODO: reorganize the layout along with adding filter
