+++ src/app/log_view/LogTab.css
@@ -24,18 +24,6 @@ body {
   padding-bottom: 200px;
 }
 
-.searchInput {
-  background-color: #f9f9f9;
-  color: #000000;
-  margin-bottom: 20px;
-  border-radius: 5px;
-  max-width: 600px;
-  padding: 15px;
-  width: 100%;
-  box-sizing: border-box;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
-}
-
 .logsContainer {
   display: flex;
   flex-direction: row;
@@ -134,6 +122,9 @@ input {
   background-color: #f9f9f9;
   box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
   box-sizing: border-box;
+  display: flex;
+  justify-content: flex-start; /* Aligns items to the left */
+  flex-wrap: wrap; /* Allows wrapping if necessary */
 }
 
 .buttonGroup {
@@ -143,6 +134,35 @@ input {
 
 .searchBarContainer {
   display: flex;
+  width: 700px;
   align-items: center;
   margin-bottom: 20px;
+  gap: 20px;
+}
+
+.searchInput {
+  flex: 2;
+  background-color: #f9f9f9;
+  color: #000000;
+  border-radius: 5px;
+  padding: 15px;
+  box-sizing: border-box;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+  width: 100%;
+}
+
+.addMedicationButton {
+  flex: 1;
+  padding: 15px;
+  border-radius: 5px;
+  background-color: #0000ff;
+  color: white;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+  align-self: stretch;
+}
+
+.addMedicationButton:hover {
+  background-color: #414141;
 }
