+++ src/app/log_view/components/LogConfirmation.css
