+++ src/app/log_view/components/LogConfirmation.tsx
